Refactor bootstrap to shrink our admin node container.
The opencrowbar/centos container was getting a little bloated.  This
pull request series refactors it to more strictly seperate package
installation from service management as well as rebasing on top of a
smaller centos base image.

The upshot is that the base admin container should be about half the
size it was.