Fix bug #157: missing DB install in development script
Correct oversight with pull #156 that tweaked the
production.sh script and docker container so that postgresql
was installed by the script (a very good thing).

Made some other improvements to ./development.sh to
setup environment and improve comments.

Tweaked documentation to match #156 behavior and prereqs.

Made some minor changes to the database.yml so that shared
attributes were more obvious.  Originally thought this was
the issue but it was not.  Changes are should be net-neutral.