Let network ranges have their own conduits.
We want to be able to create a BMC network that has an admin range to
let the admin node be issued an IP in the same subnet that the BMCs will
live on.  The easiest way to facilitate that is to make the BMC network
live on the BMC conduit by default, but have the admin range in that
network live on one of the host interfaces.