Fix how we were creating repo data.
The Chef recipes that were creating the repository information that the
rest of Crowbar was using to install packages was making some nasty
assumptions:

* It assumed that the recipe was running on the same node that the admin
  node was on.
* It assumed that it could find barclamp metadata by looking at the
  right location on the filesystem.

Both of these assumptions violated a pretty strict seperation of
concerns between the Crowbar framework and the Chef recipes.  To resolve
that, we now save all the barclamp metadata in the database at barclamp
information time, and pass that information to the provisioner via the
appropriate role sysdata.