Pull in consul for service discovery and monitoring.
Right now, we don't acutally use it for anything, but we set things up
so that the admin node is one of the servers, and the rest of the nodes
are clients.  Sadly, this does not give us HA or failover, but it is a
start.

For more information on consul, check out https://consul.io