Make nic handling logic take DHCP into better account.
On new-ish systems, dhclient gives addresses it assigns the dynamic
flag, which lets the kernel reap them once the dhclient process goes
away.  This fixes up the network recipe to take that behaviour into
account by detecting when a nic has a dhcp agent running on it and
killing the agent + flushing the nic when we take over the nic.